:truck: Move functions to orm module
