+++ pyccounting/orm/__init__.py
@@ -1,2 +1,3 @@
 from .data_model import *  # noqa: F403
 from .db import engine
+from .main import get_df, get_start_amount
